requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Flask-RESTX==1.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0